Refact

+++ Tibia.Ciclopedia.Application/UseCases/DeleteItem/DeleteItem.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Tibia.Ciclopedia.Application.BaseOutput;
-using Tibia.Ciclopedia.Domain.Interface;
+using Tibia.Ciclopedia.Domain.Items;
 
 namespace Tibia.Ciclopedia.Application.UseCases.DeleteItem
 {
@@ -19,7 +19,11 @@ namespace Tibia.Ciclopedia.Application.UseCases.DeleteItem
 
 		public async Task<Output<bool>> Handle(DeleteItemInput request, CancellationToken cancellationToken)
 		{
-			var item = await _itemRepository.GetByIdItems(request.Id, cancellationToken);
+			var item = await _itemRepository.GetItemById(request.Id, cancellationToken);
+			if (item == null)
+			{
+				return Output<bool>.Failure(new List<string> { "Não foi possível encontrar o item especificado." });
+			}
 			var result = await _itemRepository.Deletetem(item, cancellationToken);
 			return Output<bool>.Success(result);
 		}
